Change how available capacity is maintained.
Available capacity is now attached directly to the graph structure, instead
of as a function. This means that available capacity lookups are now
O(D) where D is the maximal out-degree of any node. Note that, for the balancer
application is is O(N) where N is the number of participants. Further
optimisation, could make this O(1) amortized.